// sanity.types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;

// Source: schema.json
export type FranchisePackage = {
  _id: string;
  _type: "franchisePackage";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  price?: string;
  priceNote?: string;
  features?: Array<string>;
  featured?: boolean;
  order?: number;
};

export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type Slug = {
  _type: "slug";
  current?: string;
  source?: string;
};
// ...
export type ALL_POSTS_QUERY_RESULT = Array<{
  _id: string;
  title: string | null;
  slug: Slug | null;
  mainImage: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    _type: "image";
  } | null;
  publishedAt: string | null;
  excerpt: string;
  author: {
    name: string | null;
  } | null;
  categories: Array<{
    title: string | null;
  }> | null;
}>;